// react-chat/src/types/validate.ts
export interface validateValues {
  first_name?: string;
  last_name?: string;
  username?: string;
  password?: string;
  bio?: string;
  avatar?: string;
}

export interface validateErrors {
  first_name?: string;
  last_name?: string;
  username?: string;
  password?: string;
  bio?: string;
  avatar?: string;
}